+++ packages/angular-rules/package.json
@@ -1,34 +1,24 @@
 {
   "name": "markuplint-angular-rules",
   "version": "0.1.1",
+  "type": "module",
   "license": "MIT",
   "engines": {
-    "node": ">=v10"
+    "node": ">=14"
   },
-  "main": "./lib/index.js",
-  "module": "./lib/index.es2019.mjs",
+  "main": "./lib/index.cjs",
+  "module": "./lib/index.js",
   "exports": {
     "types": "./lib/index.d.ts",
-    "import": "./lib/index.es2019.mjs",
-    "require": "./lib/index.js"
+    "import": "./lib/index.js",
+    "require": "./lib/index.cjs"
   },
   "types": "./lib/index.d.ts",
   "files": [
     "lib"
   ],
-  "scripts": {
-    "build": "tsc",
-    "test": "jest"
-  },
   "dependencies": {
-    "@markuplint/ml-core": "~2.3.4"
-  },
-  "devDependencies": {
-    "@types/jest": "~29.2.5",
-    "jest": "~29.3.1",
-    "markuplint": "~2.10.2",
-    "ts-jest": "~29.0.3",
-    "typescript": "~4.8.4"
+    "@markuplint/ml-core": "^3.7.0"
   },
   "publishConfig": {
     "access": "public"
